backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0
websockets==11.0.3
alembic==1.12.1
pytest
pytest-asyncio
